Codepart
This code is a Python script that retrieves information about laptops from a website. It uses the request library to fetch web pages, BeautifulSoup for web scraping, and pandas to create and manipulate DataFrame. The goal is to extract details such as product names, prices, and brands for laptops from multiple pages and save the information in a CSV file.